+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 NOWnews-api change logs
 
 ## [Unreleased]
+### Fixed
+- 修正 google trace engine 造成 CPU 與 Memory 消耗過大問題 @SimonSun
+
 ### Added
 - cronjob加上軍聞社新聞匯入 @appleoxxo
+- 加入 imagelab 縮圖網址與在 Image mode 加入 sizeFormat 虛擬欄位 @SimonSun
 
 ## 1.0.30 - 2017-09-04
 ### Fixed